test: test for empty items in lists, test style tweaks
