+++ test/lists/lazy-list/index.spec.js
@@ -2,4 +2,4 @@ import './instantiation.spec.js'
 import './push.spec.js'
 import './splice.spec.js'
 import './load.spec.js'
-import './to-json.spec.js'
+import './export.spec.js'
