+++ test/lists/lazy-list/instantiation.spec.js
@@ -1,5 +1,5 @@
 import assert from 'node:assert'
-import { test, before, beforeEach } from 'node:test'
+import { test, beforeEach } from 'node:test'
 
 import { LazyList } from '../../../src/list.js'
 import { Message } from '../../model/index.js'
@@ -28,7 +28,7 @@ test('LazyList', async t => {
     await t.test('passing a "type"', async t => {
       await t.beforeEach(t => {
         list = new LazyList({
-          from: [{ id: 'm_1', text: 'Hello' }, { id: 'm_2', name: 'World' }],
+          from: [{ id: 'm_1', text: 'Hello' }, { id: 'm_2', text: 'World' }],
           type: Message
         })
       })
@@ -49,9 +49,7 @@ test('LazyList', async t => {
       await t.test('throws', t => {
         assert.throws(
           () => {
-            list = new List({
-              type: User
-            })
+            list = new LazyList({ type: User })
           })
       })
     })
@@ -82,6 +80,20 @@ test('LazyList', async t => {
 
     await t.test('set to false', t => {
       assert.strictEqual(list.loaded, false)
+    })
+  })
+
+  await t.test('passing an empty items array', async t => {
+    await t.beforeEach(t => {
+      list = new LazyList({ from: [], type: Message })
+    })
+
+    await t.test('has no items', t => {
+      assert.strictEqual(list.length, 0)
+    })
+
+    await t.test('is not loaded', t => {
+      assert.ok(list.loaded === false)
     })
   })
 })
