+++ test/lists/list/index.spec.js
@@ -1,4 +1,4 @@
 import './instantiation.spec.js'
 import './push.spec.js'
 import './splice.spec.js'
-import './to-json.spec.js'
+import './export.spec.js'
