+++ test/lists/list/push.spec.js
@@ -1,8 +1,8 @@
 import assert from 'node:assert'
-import { test, before, beforeEach } from 'node:test'
+import { test, beforeEach } from 'node:test'
 
+import { Message } from '../../model/index.js'
 import { List } from '../../../src/list.js'
-import { User } from '../../model/index.js'
 
 test('List', async t => {
   let list
@@ -12,18 +12,20 @@ test('List', async t => {
 
     await t.beforeEach(t => {
       list = new List({
-        from: [{ id: 'u_1', name: 'John' }]
+        from: [{ id: 'm_1', text: 'Hello' }],
+        type: Message
       })
 
-      list.push({ id: 'u_2', name: 'Steven' })
+      list.push({ id: 'm_2', text: 'World' })
     })
 
     await t.test('adds the items', t => {
       assert.strictEqual(list.length, 2)
+      assert.strictEqual(list[0].text, 'Hello')
     })
 
     await t.test('adds them after the existing items', t => {
-      assert.strictEqual(list[1].name, 'Steven')
+      assert.strictEqual(list[1].text, 'World')
     })
   })
 })
